requirements.txt
tensorflow~=2.14.0
matplotlib~=3.8.0
numpy~=1.26.1
keras~=2.14.0
Pillow~=10.1.0